refactor: decompose GettingStartedPage (F5)
Extract the self-contained pieces of the 1081-line GettingStartedPage into
components/getting-started/: QuizSection, ChapterContent, ChapterSearch,
KeyboardHelp, and the useScrollProgress hook (used by ChapterContent).
The page drops to 634 lines and keeps its progress helpers
(getMasteryLevel/loadProgress/saveProgress) + the main orchestrator.
Pure code movement; drop the now-unused icon/hook/QUESTION_BANK imports.

Add a characterization smoke test first (the page was previously untested)
verifying the learning-center overview and module list render. 83 tests
pass, lint + build clean.
