+++ frontend/src/components/GettingStartedPage.js
@@ -1,28 +1,25 @@
 'use client';
 
-import React, { useState, useEffect, useRef, useCallback } from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
 import {
     BookOpen,
     HelpCircle,
-    Award,
-    ShieldAlert,
-    Activity,
     Search,
     CheckCircle,
-    RefreshCw,
     ChevronRight,
-    Brain,
     ChevronLeft,
     Clock,
     Target,
     BarChart3,
-    Square,
     Lock,
     Play,
-    X,
-    Keyboard
+    Keyboard,
 } from 'lucide-react';
-import { learningModules, QUESTION_BANK, getTotalChapters } from '../data/content';
+import { learningModules, getTotalChapters } from '../data/content';
+import QuizSection from './getting-started/QuizSection';
+import ChapterContent from './getting-started/ChapterContent';
+import ChapterSearch from './getting-started/ChapterSearch';
+import KeyboardHelp from './getting-started/KeyboardHelp';
 
 // --- MASTERY LEVEL HELPER ---
 const getMasteryLevel = (progress) => {
@@ -53,457 +50,6 @@ const saveProgress = (progress) => {
     }
 };
 
-// --- SCROLL PROGRESS HOOK ---
-const useScrollProgress = (ref) => {
-    const [progress, setProgress] = useState(0);
-
-    useEffect(() => {
-        const element = ref.current;
-        if (!element) return;
-
-        const handleScroll = () => {
-            const scrollTop = element.scrollTop;
-            const scrollHeight = element.scrollHeight - element.clientHeight;
-            const scrollProgress = scrollHeight > 0 ? (scrollTop / scrollHeight) * 100 : 0;
-            setProgress(Math.min(100, Math.max(0, scrollProgress)));
-        };
-
-        element.addEventListener('scroll', handleScroll, { passive: true });
-        handleScroll();
-
-        return () => element.removeEventListener('scroll', handleScroll);
-    }, [ref]);
-
-    return progress;
-};
-
-// --- QUIZ COMPONENT ---
-const QuizSection = () => {
-    const [difficulty, setDifficulty] = useState(null);
-    const [questions, setQuestions] = useState([]);
-    const [currentIndex, setCurrentIndex] = useState(0);
-    const [userAnswers, setUserAnswers] = useState({});
-    const [showResults, setShowResults] = useState(false);
-    const [score, setScore] = useState(0);
-
-    const startQuiz = (level) => {
-        setDifficulty(level);
-        const shuffled = [...QUESTION_BANK[level]].sort(() => 0.5 - Math.random());
-        setQuestions(shuffled.slice(0, 10));
-        setCurrentIndex(0);
-        setUserAnswers({});
-        setShowResults(false);
-        setScore(0);
-    };
-
-    const handleAnswer = (answer) => {
-        const currentQ = questions[currentIndex];
-        if (currentQ.type === 'checkbox') {
-            const currentSelection = userAnswers[currentQ.id] || [];
-            const newSelection = currentSelection.includes(answer)
-                ? currentSelection.filter(item => item !== answer)
-                : [...currentSelection, answer];
-            setUserAnswers({ ...userAnswers, [currentQ.id]: newSelection });
-        } else {
-            setUserAnswers({ ...userAnswers, [currentQ.id]: answer });
-        }
-    };
-
-    const nextQuestion = () => {
-        if (currentIndex < questions.length - 1) {
-            setCurrentIndex(currentIndex + 1);
-        } else {
-            calculateScore();
-            setShowResults(true);
-        }
-    };
-
-    const calculateScore = () => {
-        let newScore = 0;
-        questions.forEach(q => {
-            const userAns = userAnswers[q.id];
-            if (!userAns) return;
-            if (q.type === 'checkbox') {
-                const correct = [...q.answer].sort();
-                const user = [...userAns].sort();
-                if (JSON.stringify(correct) === JSON.stringify(user)) newScore++;
-            } else if (q.type === 'text') {
-                if (userAns.toLowerCase().trim() === q.answer.toLowerCase()) newScore++;
-            } else {
-                if (userAns === q.answer) newScore++;
-            }
-        });
-        setScore(newScore);
-    };
-
-    const resetQuiz = () => {
-        setDifficulty(null);
-        setShowResults(false);
-        setScore(0);
-    };
-
-    if (!difficulty) {
-        return (
-            <div className="flex flex-col items-center justify-center py-12 animate-in fade-in zoom-in duration-500">
-                <Brain className="w-20 h-20 text-blue-500 mb-6" />
-                <h2 className="mb-4 text-3xl font-semibold text-mm-text-primary">Test Your Market Knowledge</h2>
-                <p className="text-gray-600 dark:text-gray-400 mb-8 text-center max-w-lg">
-                    Choose a difficulty level to start a quiz. Questions cover everything from basic terms to advanced trading strategies.
-                </p>
-                <div className="grid grid-cols-1 md:grid-cols-3 gap-6 w-full max-w-2xl">
-                    {['easy', 'medium', 'hard'].map((level) => (
-                        <button
-                            key={level}
-                            onClick={() => startQuiz(level)}
-                            className={`p-6 rounded-2xl border-2 transition-all hover:-translate-y-1 hover:shadow-xl capitalize text-lg font-bold flex flex-col items-center gap-3
-                                ${level === 'easy' ? 'border-green-400 bg-green-50 dark:bg-green-900/20 text-green-700 dark:text-green-400 hover:bg-green-100' : ''}
-                                ${level === 'medium' ? 'border-yellow-400 bg-yellow-50 dark:bg-yellow-900/20 text-yellow-700 dark:text-yellow-400 hover:bg-yellow-100' : ''}
-                                ${level === 'hard' ? 'border-red-400 bg-red-50 dark:bg-red-900/20 text-red-700 dark:text-red-400 hover:bg-red-100' : ''}
-                            `}
-                        >
-                            {level === 'easy' && <CheckCircle className="w-8 h-8" />}
-                            {level === 'medium' && <Activity className="w-8 h-8" />}
-                            {level === 'hard' && <ShieldAlert className="w-8 h-8" />}
-                            {level}
-                        </button>
-                    ))}
-                </div>
-            </div>
-        );
-    }
-
-    if (showResults) {
-        const percentage = (score / questions.length) * 100;
-        let message = "Good effort!";
-        if (percentage >= 80) message = "Market Wizard! 🧙‍♂️";
-        else if (percentage >= 50) message = "Solid Knowledge! 📈";
-        else message = "Keep Learning! 📚";
-
-        return (
-            <div className="flex flex-col items-center justify-center py-12 animate-in fade-in zoom-in duration-500 text-center">
-                <Award className={`w-24 h-24 mb-6 ${percentage >= 80 ? 'text-yellow-500' : 'text-blue-500'}`} />
-                <h2 className="mb-2 text-4xl font-semibold text-mm-text-primary">Quiz Complete</h2>
-                <p className="text-xl text-gray-600 dark:text-gray-300 mb-6">{message}</p>
-                <div className="ui-panel-elevated mb-8 w-full max-w-md p-8">
-                    <div className="mb-2 text-6xl font-semibold text-mm-accent-primary">{score}/{questions.length}</div>
-                    <div className="text-sm font-bold text-gray-400 uppercase tracking-widest">Correct Answers</div>
-                </div>
-                <button onClick={resetQuiz} className="px-8 py-3 bg-gray-900 dark:bg-white text-white dark:text-gray-900 rounded-xl font-bold hover:opacity-90 transition-all flex items-center gap-2">
-                    <RefreshCw className="w-5 h-5" /> Take Another Quiz
-                </button>
-            </div>
-        );
-    }
-
-    const currentQ = questions[currentIndex];
-
-    return (
-        <div className="max-w-3xl mx-auto py-8">
-            <div className="mb-8">
-                <div className="flex justify-between text-xs font-bold uppercase text-gray-400 mb-2">
-                    <span>Question {currentIndex + 1} of {questions.length}</span>
-                    <span>{difficulty} Mode</span>
-                </div>
-                <div className="h-2 bg-gray-200 dark:bg-gray-700 rounded-full overflow-hidden">
-                    <div className="h-full bg-blue-500 transition-all duration-500" style={{ width: `${((currentIndex + 1) / questions.length) * 100}%` }} />
-                </div>
-            </div>
-            <div className="ui-panel-elevated mb-8 p-8 animate-in slide-in-from-right duration-300" key={currentQ.id}>
-                <h3 className="mb-6 text-2xl font-semibold text-mm-text-primary">{currentQ.question}</h3>
-                
-                {/* --- Multiple Choice Fix --- */}
-                {currentQ.type === 'multiple' && (
-                    <div className="space-y-3">
-                        {currentQ.options.map((option) => (
-                            <button 
-                                key={option} 
-                                onClick={() => handleAnswer(option)} 
-                                className={`w-full text-left p-4 rounded-xl border-2 transition-all font-medium ${
-                                    userAnswers[currentQ.id] === option 
-                                        ? 'border-blue-500 bg-blue-50 dark:bg-blue-900/20 text-blue-700 dark:text-blue-300' 
-                                        : 'border-gray-100 dark:border-gray-700 hover:border-blue-200 dark:hover:border-gray-600 text-gray-900 dark:text-gray-200'
-                                }`}
-                            >
-                                {option}
-                            </button>
-                        ))}
-                    </div>
-                )}
-                
-                {/* --- Checkbox Fix --- */}
-                {currentQ.type === 'checkbox' && (
-                    <div className="space-y-3">
-                        {currentQ.options.map((option) => {
-                            const isSelected = (userAnswers[currentQ.id] || []).includes(option);
-                            return (
-                                <button 
-                                    key={option} 
-                                    onClick={() => handleAnswer(option)} 
-                                    className={`w-full text-left p-4 rounded-xl border-2 transition-all font-medium flex items-center gap-3 ${
-                                        isSelected 
-                                            ? 'border-blue-500 bg-blue-50 dark:bg-blue-900/20 text-blue-700 dark:text-blue-300' 
-                                            : 'border-gray-100 dark:border-gray-700 hover:border-blue-200 dark:hover:border-gray-600 text-gray-900 dark:text-gray-200'
-                                    }`}
-                                >
-                                    <div className={`w-5 h-5 rounded border flex items-center justify-center ${isSelected ? 'bg-blue-500 border-blue-500 text-white' : 'border-gray-300 dark:border-gray-600'}`}>
-                                        {isSelected && <CheckCircle className="w-3 h-3" />}
-                                    </div>
-                                    {option}
-                                </button>
-                            );
-                        })}
-                    </div>
-                )}
-                
-                {currentQ.type === 'text' && (
-                    <input type="text" value={userAnswers[currentQ.id] || ''} onChange={(e) => handleAnswer(e.target.value)} placeholder="Type your answer here..." className="ui-input p-4 text-lg bg-transparent placeholder:text-mm-text-tertiary" />
-                )}
-            </div>
-            <div className="flex justify-end">
-                <button onClick={nextQuestion} disabled={!userAnswers[currentQ.id] || (Array.isArray(userAnswers[currentQ.id]) && userAnswers[currentQ.id].length === 0)} className="px-8 py-3 bg-blue-600 hover:bg-blue-700 text-white rounded-xl font-bold transition-all shadow-lg shadow-blue-600/20 disabled:opacity-50 disabled:cursor-not-allowed flex items-center gap-2">
-                    {currentIndex === questions.length - 1 ? 'Finish Quiz' : 'Next Question'}
-                    <ChevronRight className="w-5 h-5" />
-                </button>
-            </div>
-        </div>
-    );
-};
-
-// --- CHAPTER CONTENT COMPONENT ---
-const ChapterContent = ({ chapter, isCompleted, onToggleComplete }) => {
-    const contentRef = useRef(null);
-    const scrollProgress = useScrollProgress(contentRef);
-    const minutesLeft = Math.max(1, Math.round((chapter.estimatedMinutes || 20) * (1 - scrollProgress / 100)));
-
-    return (
-        <div className="relative">
-            {/* Reading Progress Bar */}
-            <div className="sticky top-0 z-10 h-1 bg-gray-200 dark:bg-gray-700 rounded-full overflow-hidden mb-6">
-                <div 
-                    className="h-full bg-blue-500 transition-all duration-150 ease-out"
-                    style={{ width: `${scrollProgress}%` }}
-                />
-            </div>
-
-            {/* Reading Stats */}
-            <div className="flex items-center justify-between mb-6 text-sm">
-                <div className="flex items-center gap-2 text-gray-500 dark:text-gray-400">
-                    <Clock className="w-4 h-4" />
-                    <span>{chapter.estimatedMinutes || 20} min read</span>
-                    {scrollProgress > 0 && (
-                        <>
-                            <span className="text-gray-300 dark:text-gray-600">•</span>
-                            <span className="text-blue-600 font-medium">~{minutesLeft} min left</span>
-                        </>
-                    )}
-                </div>
-                <div className="text-gray-400 dark:text-gray-500 text-xs">
-                    {Math.round(scrollProgress)}% read
-                </div>
-            </div>
-
-            {/* Content */}
-            <div 
-                ref={contentRef}
-                className="prose prose-lg max-w-none text-gray-900 dark:text-gray-100 max-h-[calc(100vh-300px)] overflow-y-auto pr-4"
-                style={{ scrollbarWidth: 'thin' }}
-            >
-                {chapter.content.map((el, index) => {
-                    switch (el.type) {
-                        case 'paragraph':
-                            return <p key={index} className="text-gray-600 dark:text-gray-300 leading-relaxed mb-4">{el.text}</p>;
-                        case 'heading':
-                            return <h3 key={index} className="text-xl font-bold text-gray-900 dark:text-white mt-8 mb-4 flex items-center gap-2">{el.text}</h3>;
-                        case 'list':
-                            return (
-                                <ul key={index} className="list-disc list-outside space-y-2 text-gray-600 dark:text-gray-300 ml-5 mb-6">
-                                    {el.items.map((li, i) => (
-                                        <li key={i} className="leading-relaxed">
-                                            {li.split(':').length > 1 ? (
-                                                <>
-                                                    <span className="font-semibold text-gray-900 dark:text-white">{li.split(':')[0]}:</span>
-                                                    {li.split(':').slice(1).join(':')}
-                                                </>
-                                            ) : li}
-                                        </li>
-                                    ))}
-                                </ul>
-                            );
-                        case 'note':
-                            return (
-                                <div key={index} className="bg-amber-50 dark:bg-amber-900/20 border-l-4 border-amber-400 p-4 my-6 rounded-r-lg">
-                                    <div className="flex items-start gap-3">
-                                        <ShieldAlert className="w-5 h-5 text-amber-600 dark:text-amber-400 shrink-0 mt-0.5" />
-                                        <p className="text-sm text-amber-800 dark:text-amber-200">{el.text}</p>
-                                    </div>
-                                </div>
-                            );
-                        default:
-                            return null;
-                    }
-                })}
-                
-                {/* Spacer for FAB */}
-                <div className="h-24" />
-            </div>
-
-            {/* Floating Action Button for Completion */}
-            <div className="fixed bottom-8 right-8 z-50">
-                <button
-                    onClick={onToggleComplete}
-                    className={`flex items-center gap-2 px-5 py-3 rounded-full shadow-lg transition-all transform hover:scale-105 ${
-                        isCompleted 
-                            ? 'bg-gray-100 text-green-600 border border-green-200 shadow-gray-200/50' 
-                            : 'bg-blue-600 text-white shadow-blue-600/30 hover:bg-blue-700'
-                    }`}
-                >
-                    {isCompleted ? <CheckCircle className="w-5 h-5" /> : <Square className="w-5 h-5" />}
-                    <span className="font-medium text-sm">
-                        {isCompleted ? 'Done' : 'Complete'}
-                    </span>
-                    {!isCompleted && <span className="ml-1 px-1.5 py-0.5 bg-white/20 rounded text-xs">M</span>}
-                </button>
-            </div>
-        </div>
-    );
-};
-
-// --- CHAPTER SEARCH COMPONENT ---
-const ChapterSearch = ({ onSelectResult, onClose }) => {
-    const [query, setQuery] = useState('');
-    const [results, setResults] = useState([]);
-    const inputRef = useRef(null);
-
-    useEffect(() => {
-        inputRef.current?.focus();
-    }, []);
-
-    useEffect(() => {
-        if (!query.trim()) {
-            setResults([]);
-            return;
-        }
-
-        const searchResults = [];
-        const lowerQuery = query.toLowerCase();
-
-        learningModules.forEach(module => {
-            module.chapters.forEach(chapter => {
-                const matchTitle = chapter.title.toLowerCase().includes(lowerQuery);
-                const matchContent = chapter.content.some(el => 
-                    el.text?.toLowerCase().includes(lowerQuery) ||
-                    el.items?.some(item => item.toLowerCase().includes(lowerQuery))
-                );
-
-                if (matchTitle || matchContent) {
-                    searchResults.push({
-                        module,
-                        chapter,
-                        type: matchTitle ? 'title' : 'content'
-                    });
-                }
-            });
-        });
-
-        setResults(searchResults.slice(0, 8));
-    }, [query]);
-
-    return (
-        <div className="fixed inset-0 z-50 flex items-start justify-center pt-32 bg-black/50 backdrop-blur-sm animate-in fade-in duration-200">
-            <div className="ui-panel-elevated w-full max-w-2xl mx-4 overflow-hidden">
-                <div className="flex items-center gap-3 p-4 border-b border-gray-200 dark:border-gray-700">
-                    <Search className="w-5 h-5 text-gray-400" />
-                    <input
-                        ref={inputRef}
-                        type="text"
-                        value={query}
-                        onChange={(e) => setQuery(e.target.value)}
-                        placeholder="Search chapters, topics, or keywords..."
-                        className="flex-1 bg-transparent text-lg outline-none text-mm-text-primary placeholder:text-mm-text-tertiary"
-                    />
-                    <button onClick={onClose} className="p-2 hover:bg-gray-100 dark:hover:bg-gray-700 rounded-lg">
-                        <X className="w-5 h-5 text-gray-400" />
-                    </button>
-                </div>
-                
-                <div className="max-h-96 overflow-y-auto">
-                    {results.length > 0 ? (
-                        results.map(({ module, chapter, type }, idx) => (
-                            <button
-                                key={`${module.id}-${chapter.id}-${idx}`}
-                                onClick={() => {
-                                    onSelectResult(module, chapter);
-                                    onClose();
-                                }}
-                                className="w-full text-left p-4 hover:bg-gray-50 dark:hover:bg-gray-700 border-b border-gray-100 dark:border-gray-700 last:border-0 transition-colors"
-                            >
-                                <div className="flex items-center gap-2 text-xs text-gray-400 mb-1">
-                                    <span>{module.title.replace(/Module \d+: /, '')}</span>
-                                    {type === 'title' && <span className="px-1.5 py-0.5 bg-blue-100 text-blue-600 rounded">Title Match</span>}
-                                </div>
-                                <h4 className="font-semibold text-gray-900 dark:text-white">{chapter.title}</h4>
-                            </button>
-                        ))
-                    ) : query.trim() ? (
-                        <div className="p-8 text-center text-gray-500">
-                            No results found for "{query}"
-                        </div>
-                    ) : (
-                        <div className="p-8 text-center text-gray-400">
-                            Type to search across all {getTotalChapters()} chapters...
-                        </div>
-                    )}
-                </div>
-
-                <div className="p-3 bg-gray-50 dark:bg-gray-900 border-t border-gray-200 dark:border-gray-700 text-xs text-gray-400 flex items-center justify-between">
-                    <span>{results.length} results</span>
-                    <div className="flex items-center gap-1">
-                        <span>Press</span>
-                        <kbd className="px-1.5 py-0.5 bg-white dark:bg-gray-800 rounded border">ESC</kbd>
-                        <span>to close</span>
-                    </div>
-                </div>
-            </div>
-        </div>
-    );
-};
-
-// --- KEYBOARD SHORTCUTS HELP ---
-const KeyboardHelp = ({ onClose }) => (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/50 backdrop-blur-sm animate-in fade-in duration-200" onClick={onClose}>
-        <div className="ui-panel-elevated w-full max-w-md mx-4 p-6" onClick={e => e.stopPropagation()}>
-            <div className="flex items-center justify-between mb-6">
-                <h3 className="flex items-center gap-2 text-lg font-semibold text-mm-text-primary">
-                    <Keyboard className="w-5 h-5" />
-                    Keyboard Shortcuts
-                </h3>
-                <button onClick={onClose} className="p-2 hover:bg-gray-100 dark:hover:bg-gray-700 rounded-lg">
-                    <X className="w-5 h-5 text-gray-400" />
-                </button>
-            </div>
-            
-            <div className="space-y-3">
-                {[
-                    { key: '← / K', desc: 'Previous chapter' },
-                    { key: '→ / J', desc: 'Next chapter' },
-                    { key: 'M', desc: 'Mark chapter complete' },
-                    { key: 'Esc', desc: 'Exit / Close modals' },
-                    { key: '/', desc: 'Open search' },
-                    { key: '?', desc: 'Show this help' },
-                ].map(({ key, desc }) => (
-                    <div key={key} className="flex items-center justify-between py-2">
-                        <span className="text-gray-600 dark:text-gray-300">{desc}</span>
-                        <kbd className="px-2 py-1 bg-gray-100 dark:bg-gray-700 rounded text-sm font-mono text-gray-700 dark:text-gray-300 shadow-sm border border-gray-200 dark:border-gray-600">
-                            {key}
-                        </kbd>
-                    </div>
-                ))}
-            </div>
-        </div>
-    </div>
-);
-
-// --- MAIN PAGE COMPONENT ---
 const GettingStartedPage = () => {
     const [activeTab, setActiveTab] = useState('learn');
     const [selectedModule, setSelectedModule] = useState(null);
